@@ -1,6 +1,6 @@
 {
   "name": "react-cardstack",
-  "version": "0.1.0",
+  "version": "0.1.2",
   "description": "react component inspired by the iOS passbook app UI",
   "main": "dist/index.js",
   "keywords": [
@@ -8,7 +8,8 @@
     "react-component",
     "carousel",
     "navigation",
-    "cards"
+    "cards",
+    "panels"
   ],
   "repository": {
     "type": "git",
@@ -16,33 +17,31 @@
   },
   "scripts": {
     "start": "webpack-dev-server --content-base example/",
-    "compile": "babel src --out-dir dist",
-		"build:umd": "webpack",
-		"build:example": "webpack --config ./example/webpack.config.js --progress --colors",
-    "prepublish": "npm run compile && npm run build:umd && npm run build:example",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build:compile": "babel src --out-dir dist",
+    "build:umd": "webpack",
+    "build:example": "webpack --config ./example/webpack.config.js --progress --colors"
   },
   "files": [
-    "dist",
-    "umd"
+    "dist"
   ],
   "author": "Cameron Bourke <cameron2295@gmail.com> (http://cameronbourke.com/)",
   "license": "MIT",
+  "dependencies": {
+    "prop-types": "^15.5.7"
+  },
   "devDependencies": {
-    "babel-cli": "^6.3.17",
-    "babel-core": "^6.3.26",
-    "babel-loader": "^6.2.0",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.3.13",
-    "babel-preset-stage-0": "^6.3.13",
-    "eslint": "^2.9.0",
-    "eslint-config-keystone": "^2.2.0",
-    "eslint-plugin-react": "^5.0.1",
-    "react-dom": "^15.0.2",
-    "webpack": "^1.12.9",
-    "webpack-dev-server": "^1.14.0"
+    "@babel/cli": "^7.0.0-beta.40",
+    "@babel/core": "^7.0.0-beta.40",
+    "@babel/plugin-proposal-object-rest-spread": "^7.0.0-beta.40",
+    "@babel/preset-env": "^7.0.0-beta.40",
+    "@babel/preset-react": "^7.0.0-beta.40",
+    "babel-loader": "8.0.0-beta.0",
+    "react": "15.3.0",
+    "react-dom": "15.3.0",
+    "webpack": "^3.11.0",
+    "webpack-dev-server": "^2.11.1"
   },
   "peerDependencies": {
-    "react": ">=0.14.5"
+    "react": ">=15.3.0"
   }
 }
